added min height/width to electron, moved modal files
