+++ src/renderer/Presentational/AddEthereumNode/AddEthereumNode.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { NodeLibrary } from 'main/state/nodeLibrary';
-import { ModalConfig } from '../../Generics/redesign/Modal/ModalManager';
+import { ModalConfig } from '../ModalManager/ModalManager';
 import {
   container,
   descriptionFont,
