+++ src/renderer/Presentational/AddNodeStepper/AddNodeStepperModal.tsx
@@ -3,7 +3,7 @@
 import { useCallback, useState } from 'react';
 
 import { SelectOption } from '../../Generics/redesign/SpecialSelect/SpecialSelect';
-import { ModalConfig } from '../../Generics/redesign/Modal/modalUtils';
+import { ModalConfig } from '../ModalManager/modalUtils';
 import ContentWithSideArt from '../../Generics/redesign/ContentWithSideArt/ContentWithSideArt';
 import { componentContainer, container } from './addNodeStepper.css';
 import AddEthereumNode, {
