+++ src/renderer/Presentational/ModalManager/modalUtils.tsx
