+++ src/renderer/Presentational/NodeSettings/NodeSettingsWrapper.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { ModalConfig } from '../../Generics/redesign/Modal/modalUtils';
+import { ModalConfig } from '../ModalManager/modalUtils';
 import { setModalState } from '../../state/modal';
 import Node, { NodeId } from '../../../common/node';
 import {
